+++ distro/config.go
@@ -46,9 +46,6 @@ const (
 	// Logging level to set when using the default logger.
 	otelLogLevelKey = "OTEL_LOG_LEVEL"
 
-	// FIXME: support OTEL_SPAN_LINK_COUNT_LIMIT
-	// FIXME: support OTEL_ATTRIBUTE_VALUE_LENGTH_LIMIT
-
 	// splunkMetricsEndpointKey defines the endpoint Splunk specific metrics
 	// are sent. This is not currently supported.
 	splunkMetricsEndpointKey = "SPLUNK_METRICS_ENDPOINT"
@@ -74,6 +71,7 @@ type exporterConfig struct {
 type config struct {
 	Logger     logr.Logger
 	Propagator propagation.TextMapPropagator
+	SpanLimits *trace.SpanLimits
 
 	ExportConfig      *exporterConfig
 	TraceExporterFunc traceExporterFunc
@@ -82,7 +80,8 @@ type config struct {
 // newConfig returns a validated config with Splunk defaults.
 func newConfig(opts ...Option) *config {
 	c := &config{
-		Logger: logger(zapConfig(envOr(otelLogLevelKey, defaultLogLevel))),
+		Logger:     logger(zapConfig(envOr(otelLogLevelKey, defaultLogLevel))),
+		SpanLimits: newSpanLimits(),
 		ExportConfig: &exporterConfig{
 			AccessToken: envOr(accessTokenKey, defaultAccessToken),
 		},